hallo

+++ python/load_mnist.py
@@ -8,17 +8,17 @@ def load_mnist(path, gpu):
 
     if gpu:
 
-        x_tr = th.from_numpy(data['a'].reshape(60000, 1, 28, 28)).cuda()
-        y_tr = th.from_numpy(data['b']).cuda()
+        x_tr = th.from_numpy(data['a'].reshape(60000, 1, 28, 28)).float().cuda()
+        y_tr = th.from_numpy(data['b']).long().cuda()
 
-        x_te = th.from_numpy(data['c'].reshape(10000, 1, 28, 28)).cuda()
-        y_te = th.from_numpy(data['d']).cuda()
+        x_te = th.from_numpy(data['c'].reshape(10000, 1, 28, 28)).float().cuda()
+        y_te = th.from_numpy(data['d']).long().cuda()
 
     else:
 
-        x_tr = th.from_numpy(data['a'].reshape(60000, 1, 28, 28))
-        y_tr = th.from_numpy(data['b'])
-        x_te = th.from_numpy(data['c'].reshape(10000, 1, 28, 28))
-        y_te = th.from_numpy(data['d'])
+        x_tr = th.from_numpy(data['a'].reshape(60000, 1, 28, 28)).float()
+        y_tr = th.from_numpy(data['b']).long()
+        x_te = th.from_numpy(data['c'].reshape(10000, 1, 28, 28)).float()
+        y_te = th.from_numpy(data['d']).long()
 
     return x_tr, y_tr, x_te, y_te
